CMS is now mostly correctl linking (including dictionaries).
This adds support for dictionary generation when building
which should solve of the missing references related to
ROOT generated IO-functions.

We now also use globbing to no longer need these long
list of SO files in the builtin.json...

Also fixes aminor bug in the stl.modulemap that came up in
the latest clang version.